refactor: move DSL operator tests into separate files
